Switching Computers

+++ lib/Stepper/Stepper.cpp
@@ -1,6 +1,6 @@
 #include <arduino.h>
 #include <Stepper.h>
 
-void cube(uint8_t amount) {
+void cube(int amount) {
     
 }
